Fixed bug which prevented use of complex TypedPipe sources.

+++ core/src/main/scala/com/rouesnel/typedsql/core/HiveType.scala
@@ -88,7 +88,7 @@ final case object TinyIntType extends PrimitiveType {
   def thriftTypeName = "BYTE"
 }
 final case object ShortType extends PrimitiveType {
-  def hiveType = "int"
+  def hiveType = "smallint"
   def scalaTypeName = "Short"
   def thriftTypeName = "I16"
 }
